Move price_raffle from resources to features
