+++ map_gen/maps/danger_ores/modules/expanse.lua
@@ -2,7 +2,7 @@ local b = require 'map_gen.shared.builders'
 local Event = require 'utils.event'
 local Generate = require 'map_gen.shared.generate'
 local Global = require 'utils.global'
-local PriceRaffle = require 'resources.price_raffle'
+local PriceRaffle = require 'features.price_raffle'
 local RS = require 'map_gen.shared.redmew_surface'
 local table = require 'utils.table'
 local Task = require 'utils.task'
