+++ map_gen/maps/frontier.lua
@@ -4,7 +4,7 @@ local Event = require 'utils.event'
 local Global = require 'utils.global'
 local math = require 'utils.math'
 local MGSP = require 'resources.map_gen_settings'
-local PriceRaffle = require 'resources.price_raffle'
+local PriceRaffle = require 'features.price_raffle'
 local RS = require 'map_gen.shared.redmew_surface'
 local ScenarioInfo = require 'features.gui.info'
 local Toast = require 'features.gui.toast'
